Fix the wrong iteration in do_shuffle()
The original code tail = tail->prev in line 801 will cause the algorithm to move
tail pointer to node ahead than it should be since tail pointer has changed it's
place in the list.

Use a tmp pointer to record tail->prev and replace the wrong code with tail = tmp
to fix the problem.